requirements.txt
build==0.7.0
click==8.1.3
requests==2.27.1
pyyaml==6.0
